Feat: Created Laboratory handler.

+++ internal/services/laboratory_service.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/CABGenOrg/cabgen_backend/internal/models"
 	"github.com/CABGenOrg/cabgen_backend/internal/repository"
+	"github.com/CABGenOrg/cabgen_backend/internal/validations"
 	"github.com/google/uuid"
 	"gorm.io/gorm"
 )
@@ -16,8 +17,8 @@ type LaboratoryService interface {
 	FindByID(ctx context.Context, ID uuid.UUID) (*models.Laboratory, error)
 	FindByNameOrAbbreviation(ctx context.Context, input string) ([]models.Laboratory, error)
 	Create(ctx context.Context, lab *models.Laboratory) error
-	Update(ctx context.Context, lab *models.Laboratory) error
-	Delete(ctx context.Context, lab *models.Laboratory) error
+	Update(ctx context.Context, ID uuid.UUID, input models.LaboratoryUpdateInput) (*models.Laboratory, error)
+	Delete(ctx context.Context, ID uuid.UUID) error
 }
 
 type laboratoryService struct {
@@ -89,24 +90,42 @@ func (s *laboratoryService) Create(ctx context.Context, lab *models.Laboratory) 
 	return nil
 }
 
-func (s *laboratoryService) Update(ctx context.Context, lab *models.Laboratory) error {
-	existingLab, err := s.Repo.GetLaboratoryDuplicate(ctx, lab.Name, lab.ID)
+func (s *laboratoryService) Update(ctx context.Context, ID uuid.UUID, input models.LaboratoryUpdateInput) (*models.Laboratory, error) {
+	existingLab, err := s.Repo.GetLaboratoryByID(ctx, ID)
+	if errors.Is(err, gorm.ErrRecordNotFound) {
+		return nil, ErrNotFound
+	}
+	if err != nil {
+		return nil, ErrInternal
+	}
+
+	validations.ApplyLaboratoryUpdate(existingLab, &input)
+
+	duplicate, err := s.Repo.GetLaboratoryDuplicate(ctx, existingLab.Name, existingLab.ID)
 	if err != nil && !errors.Is(err, gorm.ErrRecordNotFound) {
-		return ErrInternal
+		return nil, ErrInternal
 	}
 
-	if existingLab != nil {
-		return ErrConflict
+	if duplicate != nil {
+		return nil, ErrConflict
 	}
 
-	if err := s.Repo.UpdateLaboratory(ctx, lab); err != nil {
-		return ErrInternal
+	if err := s.Repo.UpdateLaboratory(ctx, existingLab); err != nil {
+		return nil, ErrInternal
 	}
 
-	return nil
+	return existingLab, nil
 }
 
-func (s *laboratoryService) Delete(ctx context.Context, lab *models.Laboratory) error {
+func (s *laboratoryService) Delete(ctx context.Context, ID uuid.UUID) error {
+	lab, err := s.Repo.GetLaboratoryByID(ctx, ID)
+	if errors.Is(err, gorm.ErrRecordNotFound) {
+		return ErrNotFound
+	}
+	if err != nil {
+		return ErrInternal
+	}
+
 	if err := s.Repo.DeleteLaboratory(ctx, lab); err != nil {
 		return ErrInternal
 	}
